Deinterlace interlaced video.
When an interlaced video is detected, pass it through the bwdif filter to deinterlace it prior to encoding.
Added ENTRYPOINT to the Dockerfile so that when executing via docker run we do not need to explicitly call compress-mkv.
Only probe a file one time as it can be a costly operation to duplicate when working with large (4K UHD) files.
Don't error out when probing fails; skip to the next file instead.